+++ CHANGELOG.md
@@ -45,6 +45,7 @@ FEATURES:
 * **New Resource:** `conformity_aws_account`
 * **New Resource:** `conformity_azure_account`
 * **New Resource:** `conformity_gcp_account`
+* **New Resource:** `conformity_gcp_org`
 * **New Resource:** `conformity_group`
 * **New Resource:** `conformity_user`
 * **New Resource:** `conformity_sso_user`